Renamed 1.7 and 1.7.10 to 1.7-R4
as it is the only supported 1.7.X version at the moment.
