+++ src/main/java/co/marcin/novaguilds/manager/ConfigManager.java
@@ -62,7 +62,7 @@ public class ConfigManager {
 
 	public enum ServerVersion {
 		MINECRAFT_1_7_2,
-		MINECRAFT_1_7_10,
+		MINECRAFT_1_7_R4,
 		MINECRAFT_1_8,
 		MINECRAFT_1_9_R1,
 		MINECRAFT_1_9_R2,
