+++ src/main/java/co/marcin/novaguilds/manager/MessageManager.java
@@ -262,7 +262,7 @@ public class MessageManager {
 
 		switch(ConfigManager.getServerVersion()) {
 			case MINECRAFT_1_7_2:
-			case MINECRAFT_1_7_10:
+			case MINECRAFT_1_7_R4:
 				throw new IllegalArgumentException("Cannot send a title on 1.7 server");
 			case MINECRAFT_1_8:
 				title = new co.marcin.novaguilds.impl.versionimpl.v1_8.TitleImpl();
